always honor `GLOMMIO_TEST_POLLIO_ROOTDIR` when testing
`ImmutableFile` unit tests were not using the value provided via the
envvar `GLOMMIO_TEST_POLLIO_ROOTDIR`.

Also, allow users to provide a comma-separated list of path instead of a
single one. That's useful when testing on different filesystems.
